intergrated traverse into main Cha code removing old \ls logic
